refactor: derive the Flix runtime class names from one descriptor
ExternalJarLoader spelled dev.flix.runtime.Global twice, once as a
binary name and once as a class file path, and the dev.flix.test package
twice in the same two forms. The Global descriptor already existed as
GenGlobal.Desc, so the loader held a third and fourth spelling of names
the backend already knew.

Add FlixClasses, alongside JavaClasses, holding the two Flix packages and
the Global descriptor. GenGlobal and Mangle.DevFlixRuntime now derive from
it, and the loader derives its four strings through the existing ClassDescs
helpers rather than repeating them as literals.

Move ExternalJarLoader from util into language.jvm so it can reach those
constants. It is entirely a JVM class-loading concern, and language.jvm is
where the JVM naming and metadata layer already lives.
